requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
pillow
langchain-community
bs4
faiss-cpu
pypdf
arxiv
pymupdf
wikipedia
lxml
langchain_huggingface
langchain-groq
langgraph>=0.0.16
langgraph-cli[inmem]
youtube_transcript_api
langchain-text-splitters
langchain_chroma
langchain_huggingface
langchain_openai
langchain_community
langchain_groq
openai_embeddings
ollama_embeddings
huggingface_embeddings
gtts
typing-extensions
# Optional debug visualization dependencies
websockets
